+++ packages/wallet-connectors/src/WalletConnection.ts
@@ -3,10 +3,12 @@ import { SendTransactionPayload, SmartContractParameters } from '@concordium/bro
 import {
     AccountTransactionSignature,
     AccountTransactionType,
+    ConcordiumGRPCClient,
     JsonRpcClient,
     SchemaVersion,
     toBuffer,
 } from '@concordium/web-sdk';
+import { GrpcWebOptions } from '@protobuf-ts/grpcweb-transport';
 
 export type ModuleSchema = {
     type: 'ModuleSchema';
@@ -152,6 +154,7 @@ export interface WalletConnection {
     /**
      * Returns a JSON-RPC client that is ready to perform requests against some Concordium Node connected to network/chain
      * that the connected account lives on.
+     * The client implements version 1 of the Node's API which is deprecated. Prefer using {@link getGrpcClient} instead.
      *
      * This method is included because it's part of the Browser Wallet's API.
      * It should be used with care as it's hard to guarantee that it actually connects to the expected network.
@@ -160,10 +163,28 @@ export interface WalletConnection {
      * Note that this method cannot be moved to {@link WalletConnector} as the Browser Wallet's RPC client doesn't work
      * until a connection has been established.
      *
-     * @return Returns a JSON-RPC client for performing requests against a Concordium Node connected
-     * to the appropriate network.
+     * @return A JSON-RPC client for performing requests against a Concordium Node connected to the appropriate network.
+     * @throws If the connection uses {@link Network.jsonRpcUrl} and that value is undefined.
+     * @deprecated Use {@link getGrpcClient} instead.
      */
     getJsonRpcClient(): JsonRpcClient;
+
+    /**
+     * Returns a gRPC client that is ready to perform requests against some Concordium Node connected to network/chain
+     * that the connected account lives on.
+     * The client implements version 2 of the Node's API and is preferred over {@link getJsonRpcClient}.
+     *
+     * This method is included because it's part of the Browser Wallet's API.
+     * It should be used with care as it's hard to guarantee that it actually connects to the expected network.
+     * The application may easily instantiate its own client and use that instead for more control.
+     *
+     * Note that this method cannot be moved to {@link WalletConnector} as the Browser Wallet's RPC client doesn't work
+     * until a connection has been established.
+     *
+     * @return A gRPC client for performing requests against a Concordium Node connected to the appropriate network.
+     * @throws If the connection uses {@link Network.grpcOpts} and that value is undefined.
+     */
+    getGrpcClient(): ConcordiumGRPCClient;
 
     /**
      * Assembles a transaction and sends it off to the wallet for approval and submission.
@@ -227,10 +248,51 @@ export interface Network {
     genesisHash: string;
 
     /**
+     * The connection configuration for a gRPC Web endpoint for performing API (v2) queries against a
+     * Concordium Node instance connected to this network.
+     *
+     * The value is currently used only for {@link WalletConnectConnection WalletConnect connections}
+     * as {@link BrowserWalletConnector Browser Wallet connections} use the Browser Wallet's internal client.
+     *
+     * Making the field undefined disables the automatic initialization of the client for WalletConnect connections.
+     * Instead of returning a client instance, {@link WalletConnectConnection.getGrpcClient} will throw an exception
+     * in this case.
+     *
+     * There's no fundamental difference between using a manually configured client compared to the one belonging to the connections.
+     * Keeping it separate from connection puts the dApp in control and also allows it to use the client
+     * before any connections have been established.
+     * The initialization is straightforward:
+     * <pre>
+     *   import { ConcordiumGRPCClient } from '@concordium/web-sdk';
+     *   import { GrpcWebFetchTransport } from '@protobuf-ts/grpcweb-transport';
+     *   ...
+     *   const client = new ConcordiumGRPCClient(new GrpcWebFetchTransport(network.grpcOpts!));
+     * </pre>
+     */
+    grpcOpts: GrpcWebOptions | undefined;
+
+    /**
      * The URL of a <a href="https://github.com/Concordium/concordium-json-rpc">Concordium JSON-RPC proxy</a> instance
      * for performing API (v1) queries against a Concordium Node instance connected to this network.
+     *
+     * The value is currently used only for {@link WalletConnectConnection WalletConnect connections}
+     * as {@link BrowserWalletConnector Browser Wallet connections} use the Browser Wallet's internal client.
+     *
+     * Making the field undefined disables the automatic initialization of the client for WalletConnect connections.
+     * Instead of returning a client instance, {@link WalletConnectConnection.getJsonRpcClient} will throw an exception
+     * in this case.
+     *
+     * There's no fundamental difference between using a manually configured client compared to the one belonging to the connections.
+     * Keeping it separate from connection puts the dApp in control and also allows it to use the client
+     * before any connections have been established.
+     * The initialization is straightforward:
+     * <pre>
+     *   import { HttpProvider, JsonRpcClient } from '@concordium/web-sdk';
+     *   ...
+     *   const client = new JsonRpcClient(new HttpProvider(network.jsonRpcUrl!));
+     * </pre>
      */
-    jsonRpcUrl: string;
+    jsonRpcUrl: string | undefined;
 
     /**
      * The base URL of a <a href="https://github.com/Concordium/concordium-scan">CCDScan</a> instance